+++ Cornucopia/Resampler.cpp
@@ -196,7 +196,7 @@ public:
 
     int paramToIdx(double param, double *outParam) const
     {
-        int idx = min(std::upper_bound(_lengths.begin(), _lengths.end(), param) - _lengths.begin(), (ptrdiff_t)_lengths.size() - 1) - 1;
+        int idx = (int)min(std::upper_bound(_lengths.begin(), _lengths.end(), param) - _lengths.begin(), (ptrdiff_t)_lengths.size() - 1) - 1;
         if(outParam)
             *outParam = param - _lengths[idx];
         return idx;
